refact: refactor project
